third_party/blink/web_tests/http/tests/devtools/sources/debugger-breakpoints/xhr-breakpoints-expected.txt

Tests XHR breakpoints.


Running: testXHRBreakpoint
Script execution paused.
Call stack:
    0) sendRequest (xhr-breakpoints.js:18)
    1)  (:1)
Script execution resumed.

Running: testPauseOnAnyXHR
Script execution paused.
Script execution resumed.
Script execution paused.
Script execution resumed.

Running: testDisableBreakpoint
Script execution paused.
Script execution resumed.
